pyproject.toml
[tool.poetry]
name = "bruteforcer"
version = "0.1.0"
description = ""
authors = ["AiWaldoh <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pymysql = "^1.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
